+++ NEWS.md
@@ -1,4 +1,11 @@
 # tbgeneratr 0.1.1.2
+#### 9 Apr 2019
+## Bug fix
+* fix class detection 
+
+---
+
+# tbgeneratr 0.1.1.2
 #### 25 Feb 2019
 ## Minor update
 * `age_generator()`: add `categorise` argument to generate additional factor variable for age.